+++ changelog.md
@@ -5,6 +5,7 @@
 - With this new technology comes:
     - Support for up to 32 bots per team (previously dependant on the selected 'rlbot.cfg'). #78
     - The user no longer has to select a 'rlbot.cfg'
+    - Ability to set game mode, map and mutators under the RLBot Settings tab
 
 
 #### Version 1.3 - 9. April 2019